* Cookie case moved to fake spec
* Adapted cookie resolving case for kotlin to a separate use-case in fake spec file
* Adapted case config to take the new fake spec file

+++ samples/client/others/rust/hyper/oneOf/src/models/pizza.rs
@@ -14,7 +14,7 @@
 #[derive(Clone, Default, Debug, PartialEq, Serialize, Deserialize)]
 pub struct Pizza {
     #[serde(rename = "pizzaSize", skip_serializing_if = "Option::is_none")]
-    pub pizza_size: Option<f32>,
+    pub pizza_size: Option<f64>,
     /// Hyperlink reference
     #[serde(rename = "href", skip_serializing_if = "Option::is_none")]
     pub href: Option<String>,
